Realtime-API-Kafka_spark-main/requirements.txt
kafka-python==2.3.0
requests==2.32.5
pyspark==4.1.1
streamlit==1.41.1
pandas==2.2.3
plotly==5.24.1
